Completed insert test
