--- app/src/test/java/com/example/elevatewebsolutions_tasktracker/database/TaskDAOTest.java
@@ -1,41 +1,0 @@
-package com.example.elevatewebsolutions_tasktracker.database;
-
-import junit.framework.TestCase;
-
-public class TaskDAOTest extends TestCase {
-
-    public void setUp() throws Exception {
-        super.setUp();
-    }
-
-    public void tearDown() throws Exception {
-        super.tearDown();
-    }
-
-    public void testInsert() {
-    }
-
-    public void testUpdate() {
-    }
-
-    public void testDelete() {
-    }
-
-    public void testDeleteAllTasks() {
-    }
-
-    public void testGetAllTasks() {
-    }
-
-    public void testGetTasksByUserId() {
-    }
-
-    public void testGetTaskById() {
-    }
-
-    public void testGetTasksByStatus() {
-    }
-
-    public void testSearchTasks() {
-    }
-}
